msm: ipa: have gsi_alloc_evt_ring() return id
Instead of passing the address of an object to receive the allocated
ring id, just have gsi_alloc_evt_ring() return the allocated id, or
a negative errno if an error occurs.  Update the comments for this
function and gsi_dealloc_evt_ring() to reflect that they don't
return a gsi_status any more (this should have been done earlier),

+++ drivers/platform/msm/ipa_lite/ipa_i.h
@@ -1815,7 +1815,7 @@ void ipa3_reset_freeze_vote(void);
 struct dentry *ipa_debugfs_get_root(void);
 bool ipa3_is_msm_device(void);
 void ipa3_enable_dcd(void);
-int ipa3_alloc_common_event_ring(void);
+long ipa3_alloc_common_event_ring(void);
 int ipa3_allocate_dma_task_for_gsi(void);
 void ipa3_free_dma_task_for_gsi(void);
 int ipa3_disable_apps_wan_cons_deaggr(uint32_t agg_size, uint32_t agg_count);
